implement order status actions

+++ app/Http/Controllers/WashOrderController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\OrderStatus;
 use Spatie\QueryBuilder\QueryBuilder;
 
 use App\Models\WashOrder;
@@ -63,6 +64,10 @@ class WashOrderController extends Controller
      */
     public function update(UpdateWashOrderRequest $request, WashOrder $order)
     {
+        if ($order->status !== OrderStatus::CREATED->value) {
+            return $this->respondForbidden();
+        }
+
         $order->update($request->all());
 
         foreach ($order->details as $detail) {
@@ -80,10 +85,31 @@ class WashOrderController extends Controller
      */
     public function destroy(WashOrder $order)
     {
+        if ($order->status !== OrderStatus::CREATED->value) {
+            return $this->respondForbidden();
+        }
+
         $order->delete();
 
         return $this->respondWithSuccess([
             'message' => 'Wash Order has been deleted'
+        ]);
+    }
+
+    /**
+     * Approve the specified order
+     */
+    public function approveOrder(WashOrder $washOrder)
+    {
+        if ($washOrder->status !== OrderStatus::CREATED->value) {
+            return $this->respondError('Wash Order cannot be approved');
+        }
+
+        $washOrder->status = OrderStatus::APPROVED->value;
+        $washOrder->saveOrFail();
+
+        return $this->respondWithSuccess([
+            'message' => 'Wash Order has been approved'
         ]);
     }
 }
